Detection improvements
Makes two versions of detection: one optimized for speed (Ba_detection_Alice_DMA_BrightDetectionOnly.py), which records the pulses for only the bright detection and then plays that back as quickly as possible. This is not currently configured for scanning. The other detection option (Ba_detection_Alice_DMA_complete.py) still uses DMA, so is faster than previous versions, but does both detections (sigma 1 and sigma 2) and works for all scans. It is therefore somewhat slower, although it is unclear how much. DDS for Alice 493 sigma 2 is also switched.